修改客显抓取订单的方式

+++ app/src/main/java/com/gt/magicbox/http/ApiService.java
@@ -83,7 +83,7 @@ public interface ApiService {
             , @Query("current") int current, @Query("size") int size);
 
     @POST(HttpConfig.DELETE_NOT_PAY_ORDER)
-    Observable<BaseResponse> deleteNotPayOrder(@Query("eqId") int eqId, @Query("orderId") int orderId);
+    Observable<BaseResponse> deleteNotPayOrder(@Query("eqId") int eqId, @Query("orderId") long orderId);
 
     @GET("magicBoxMobile/{orderId}/{shiftId}/{busId}//79B4DE7C/payQR")
     Observable<BaseResponse<CreatedOrderBean>> getCreatedQRCodeUrl(@Path("orderId") int orderId, @Path("shiftId") int shiftId,
